multiparameter writing working?
