+++ run.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+./build.sh
+./compiler test.portugol
+clang -o prog module.bc
+./prog
