+++ src/parser.y
@@ -126,7 +126,7 @@ int yyerror();
 %type <node> COMPARISON_OPERATOR SWITCH_STATEMENT SWITCH_HEADER SWITCH_DECLARE SWITCH_BODY
 %type <node> CASE_BODY CASE_LIST SWITCH_CASE CASE_BEGIN CASE_DECLARE EXTRA_CASES
 %type <node> EXTRA_CASE_LIST EXTRA_CASE_DECLARE DEFAULT_CASE DEFAULT_HEAD
-%type <node> WRITE_STATEMENT WRITE_BEGIN WRITE_END WRITE_PARAMS EXTRA_PARAMS
+%type <node> WRITE_STATEMENT WRITE_BEGIN WRITE_END WRITE_PARAMS EXTRA_PARAMS EXTRA_PARAM
 %type <node> READ_STATEMENT READ_BEGIN READ_END READ_PARAMS
 %type <node> FOR_STATEMENT FOR_HEADER FOR_DECLARE_VAR FOR_DECLARE_BODY RANGE_DECLARE
 %type <node> START_RANGE END_RANGE STEP_DECLARE DO_DECLARE FOR_BODY FOR_END
@@ -267,11 +267,12 @@ WRITE_BEGIN: WRITE_ OPEN_CURLY_ { $$ = createNode(TOKEN, (NodeValue) "WRITE_BEGI
 WRITE_END: WRITE_PARAMS CLOSE_CURLY_ { $$ = createNode(TOKEN, (NodeValue) "WRITE_END", $1, $2); };
 WRITE_PARAMS: VAR_NAME_  { $$ = createNode(TOKEN, (NodeValue) "WRITE_PARAMS", $1, NULL); }
             | STRING_LIT_  { $$ = createNode(TOKEN, (NodeValue) "WRITE_PARAMS", $1, NULL); }
-            | VAR_NAME_ EXTRA_PARAMS {}
-            | STRING_LIT_ EXTRA_PARAMS {};
-EXTRA_PARAMS: EXTRA_PARAM  {}
-            | EXTRA_PARAMS EXTRA_PARAM {};
-EXTRA_PARAM: COMMA_ STRING_LIT_ | COMMA_ VAR_NAME_ {};
+            | VAR_NAME_ EXTRA_PARAMS { $$ = createNode(TOKEN, (NodeValue) "WRITE_PARAMS", $1, $2); }
+            | STRING_LIT_ EXTRA_PARAMS { $$ = createNode(TOKEN, (NodeValue) "WRITE_PARAMS", $1, $2); };
+EXTRA_PARAMS: EXTRA_PARAM  { $$ = createNode(TOKEN, (NodeValue) "EXTRA_PARAMS", $1, NULL); }
+            | EXTRA_PARAM EXTRA_PARAMS { $$ = createNode(TOKEN, (NodeValue) "EXTRA_PARAMS", $1, $2); };
+EXTRA_PARAM: COMMA_ STRING_LIT_ { $$ = createNode(TOKEN, (NodeValue) "EXTRA_PARAM", $1, $2); };
+            | COMMA_ VAR_NAME_ { $$ = createNode(TOKEN, (NodeValue) "EXTRA_PARAM", $1, $2); };
 
 READ_STATEMENT: READ_BEGIN READ_END;
 READ_BEGIN: READ_ OPEN_CURLY_;
